Added fix from DRM on sync issue with >1.35 watch firmware
NOTE: from version 1.50 onwards of the watch firmware a ugly status bar
is displayed on all application pages - including the analog face app:(
It was not displayed on firmware 1.35 nor 1.40. It is no big issue to
remove (one line to comment out). The gen2.txt is in the watch download
directory. I can't generate a a43...